+++ Common/include/tool/ProtobufDispatcher.h
@@ -17,7 +17,7 @@ class Callback
 {
 public:
     virtual ~Callback() = default;
-    virtual void onMessage(const uint32_t,
+    virtual void onMessage(const uint64_t,
         const MessagePtr&,
         const int64_t&) const = 0;
 };
@@ -28,13 +28,13 @@ class CallbackT : public Callback
     static_assert(std::is_base_of<google::protobuf::Message, T>::value,
         "T must be derived from gpb::Message.");
 public:
-    using ProtobufMessageTCallback = std::function<void(const uint32_t, const std::shared_ptr<T>& message, const int64_t&)>;
+    using ProtobufMessageTCallback = std::function<void(const uint64_t, const std::shared_ptr<T>& message, const int64_t&)>;
 
     CallbackT(const ProtobufMessageTCallback& callback_)
         : m_callback(callback_) {
     }
 
-    void onMessage(const uint32_t socket_,
+    void onMessage(const uint64_t socket_,
         const MessagePtr& msg_,
         const int64_t& receive_time_) const override {
         std::shared_ptr<T> _concrete = std::static_pointer_cast<T>(msg_);
@@ -48,13 +48,13 @@ private:
 
 class ProtobufDispatch : public Singleton<ProtobufDispatch> {
 public:
-    using ProtobufMessageCallback = std::function<void(const uint32_t, const MessagePtr&, const int64_t&)>;
+    using ProtobufMessageCallback = std::function<void(const uint64_t, const MessagePtr&, const int64_t&)>;
 
-    ProtobufDispatch() :m_defaultCallback([](const uint32_t, const MessagePtr& msg_, const int64_t&) {
+    ProtobufDispatch() :m_defaultCallback([](const uint64_t, const MessagePtr& msg_, const int64_t&) {
         LogInfo << "[ProtobufDispatch] onUnknowMessageType: " << msg_->GetTypeName() << FlushLog;
     }) {}
 
-    void onProtobufMessage(const uint32_t conn_,
+    void onProtobufMessage(const uint64_t conn_,
         const MessagePtr& message_,
         const int64_t& receive_time_) const {
         CallbackListMap::const_iterator it = m_callbacks.find(message_->GetDescriptor());
